lenovo/pyproject.toml
[project]
name = "agent-ops"
version = "1.0.0"
description = "Self-hosted Agent Ops: PR-only software agent execution + audit + queue"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.128.0",
  "uvicorn==0.40.0",
  "pydantic==2.12.5",
  "pydantic-settings==2.10.1",
  "sqlalchemy==2.0.45",
  "psycopg[binary]==3.2.9",
  "redis==7.1.0",
  "rq==2.6.1",
  "httpx==0.28.1",
  "paramiko==4.0.0",
  "python-dotenv==1.0.1",
  "structlog==24.4.0",
  "tenacity==9.0.0",
  # Optional (real "software agent" SDK). Not used by tests.
  "openhands-sdk==0.4.2",
  "openhands-tools==0.4.2",
]
[project.optional-dependencies]
dev = [
  "pytest==8.3.4",
  "pytest-asyncio==0.25.2",
  "requests==2.32.3",
  "ruff==0.8.5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py312"
